requirements.txt
python-dateutil==1.5
evernote==1.25.0
colorama==0.2.5
